rename: 파일 위치변경
Provider 파일 components폴더 -> context폴더
