+++ src/components/atoms/card-expired-period/ExpiredPeriodInputContainer.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useAppDispatch, useAppState } from "../../../hooks/hooks";
-import { createAction } from "../../Provider";
+import { createAction } from "../../../context/Provider";
 import { ActionType } from "../../../types";
 import { isNum, removeWhiteSpaces, removeSlash, transformToMMYY } from "../../../utils";
 import ExpiredPeriodInput from "./ExpiredPeriodInput";
